requirements.txt
Flask==2.0.1
PyYAML==5.4.1
gunicorn==20.1.0
Werkzeug==2.2.2
google-cloud-logging